refactor(server-portal): use a main.ts for easier switch between frameworks
